pipedream-connect-examples/managed-auth-basic-next-app/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
backoff==2.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
FalkorDB==1.2.0
fastapi==0.115.0
graphiti-core==0.20.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
neo4j==5.28.2
numpy==2.3.3
openai==1.109.1
packaging==25.0
pipedream==1.0.8
postgrest==2.20.0
posthog==6.7.6
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.6.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.20.0
redis==5.3.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
